Custom hook in invoice to add line item
